# packages/contracts/openapi/aura-note.v1.yaml
openapi: 3.1.0
info:
  title: AURA Note v1 API
  version: 0.1.0
  description: Initial API contract seed for AURA Note standalone and ClinicOS-integrated modes.
servers:
  - url: /api/v1
paths:
...
  /schedule/appointments:
    get:
      operationId: listAppointments
      parameters:
        - name: activeDate
          in: query
          required: false
          schema: { type: string, format: date }
        - name: viewMode
          in: query
          required: false
          schema: { type: string, enum: [day, week] }
        - name: providerId
          in: query
          required: false
          schema: { type: string }
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [scheduled, checked_in, in_room, visit_started, visit_paused, visit_completed, finalization_in_progress, finalized, cancelled, no_show] }
        - name: visitType
          in: query
          required: false
          schema: { type: string }
        - name: modality
          in: query
          required: false
          schema: { type: string, enum: [in_person, telehealth, phone] }
        - name: clinicLocationId
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Appointment list with note shell status, backend filters, room/location metadata, and metadata-only chart-intake state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleViewResponse'
    post:
      operationId: createAppointment
      description: Create standalone appointment and one-to-one note shell.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
      responses:
        '201':
          description: Created appointment and one-to-one note shell.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAppointmentResponse'
  /standalone/patients:
    get:
      operationId: searchStandalonePatients
      description: WO-038 implemented synthetic standalone patient shell search. Real PHI and production patient matching are out of scope.
      responses:
        '200':
          description: Synthetic standalone patient shells.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandalonePatientSearchResponse'
    post:
      operationId: createStandalonePatientShell
      description: Create a standalone patient shell with a safe synthetic patient identifier.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStandalonePatientRequest'
      responses:
        '201':
          description: Standalone patient shell created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandalonePatientResponse'
  /standalone/patients/{safePatientId}:
    patch:
      operationId: updateStandalonePatientShell
      description: Update synthetic standalone patient shell metadata without real PHI.
      parameters:
        - name: safePatientId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStandalonePatientRequest'
      responses:
        '200':
          description: Standalone patient shell updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandalonePatientResponse'
...
  /standalone/operations/tasks/{taskId}:
    patch:
      operationId: updateStandaloneOperationTask
      description: Update task adjudication and blocker status with audit/event evidence.
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOperationalTaskRequest'
      responses:
        '200':
          description: Task updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalActionResponse'
...
  /standalone/operations/billing-review/{billingReviewId}:
    patch:
      operationId: updateStandaloneBillingReview
      description: Update billing review status without submitting claims.
      parameters:
        - name: billingReviewId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBillingReviewRequest'
      responses:
        '200':
          description: Billing review updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalActionResponse'
...
  /standalone/operations/settings/integrations/{integrationId}:
    patch:
      operationId: updateStandaloneIntegrationStatus
      description: Update disabled/mock integration scaffold status; no live credentials are accepted.
      parameters:
        - name: integrationId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIntegrationRequest'
      responses:
        '200':
          description: Integration status updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalActionResponse'
  /standalone/operations/templates:
    get:
      operationId: listStandaloneTemplates
      description: List synthetic templates and dot phrases with safe variables only.
      responses:
        '200':
          description: Templates and dot phrases.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplatesResponse'
    post:
      operationId: createStandaloneTemplate
      description: Create a synthetic template; obvious PHI and unsafe variables are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Template created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalActionResponse'
  /standalone/operations/dot-phrases/{dotPhraseId}:
    patch:
      operationId: updateStandaloneDotPhrase
      description: Update a synthetic dot phrase; obvious PHI and unsafe variables are rejected.
      parameters:
        - name: dotPhraseId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDotPhraseRequest'
      responses:
        '200':
          description: Dot phrase updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalActionResponse'
  /standalone/operations/estimate-config:
    get:
      operationId: getStandaloneEstimateConfig
      description: View internal-only estimate configuration; patient-facing estimates are disabled.
      responses:
        '200':
          description: Estimate configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateConfigurationResponse'
    patch:
      operationId: updateStandaloneEstimateConfig
      description: Update internal-only estimate caveats; patient-facing financial conclusions are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEstimateConfigurationRequest'
      responses:
        '200':
          description: Estimate configuration updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalActionResponse'
...
  /standalone/operations/rules-catalog/publish:
    post:
      operationId: publishStandaloneRulesCatalogEntries
      description: Publish synthetic rules only with human-review attestation; no autonomous coding or medical-necessity decision is enabled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRulesCatalogRequest'
      responses:
        '201':
          description: Rules catalog entries published.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalActionResponse'
  /schedule/appointments/{appointmentId}:
    patch:
      operationId: updateStandaloneAppointment
      description: Edit appointment metadata while preserving the one appointment to one note invariant.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppointmentRequest'
      responses:
        '200':
          description: Appointment updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentActionResponse'
  /schedule/appointments/{appointmentId}/status:
    post:
      operationId: updateStandaloneAppointmentStatus
      description: Check in, cancel, or mark no-show for a standalone appointment.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentStatusActionRequest'
      responses:
        '201':
          description: Appointment lifecycle status updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentActionResponse'
  /schedule/appointments/{appointmentId}/chart-context:
    get:
      operationId: getStandaloneChartContextSnapshot
      description: Return source-linked synthetic/local chart-context snapshot metadata; live EHR completeness is not implied.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Synthetic standalone chart-context snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandaloneChartContextResponse'
  /schedule/appointments/{appointmentId}/workspace-validation:
    get:
      operationId: validateWorkspaceEntry
      description: Validate appointment, note, and chart-context linkage before opening the documentation workspace. This returns metadata-only safe identifiers and does not imply live EHR completeness.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Workspace validation metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceValidationResponse'
  /schedule/appointments/{appointmentId}/chart-intake-status:
    post:
      operationId: updateChartIntakeStatus
      description: Update metadata-only chart-intake status for the schedule/workspace handoff. Live PHI file upload and patient portal delivery remain disabled.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChartIntakeStatusRequest'
      responses:
        '201':
          description: Metadata-only chart-intake status updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartIntakeStatusResponse'
  /schedule/appointments/{appointmentId}/start-visit:
    post:
      operationId: startVisit
      description: Starts visit session scaffold and activates note shell. Full timer, recording, and transcription depth is implemented in later work orders.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        '201':
          description: Visit started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartVisitResponse'
...
components:
  schemas:
    ApiMeta:
      type: object
      properties:
        requestId: { type: string }
        traceId: { type: string }
        mode: { type: string, enum: [standalone, clinicos_integrated] }
        generatedAt: { type: string, format: date-time }
...
    Appointment:
      type: object
      required: [appointmentId, tenantId, siteId, safePatientId, clinicianId, noteId, state, startsAt, durationMinutes, visitType, modality, source, mode]
      properties:
        appointmentId: { type: string }
        tenantId: { type: string }
        siteId: { type: string }
        safePatientId: { type: string }
        clinicianId: { type: string }
        noteId: { type: string }
        state:
          type: string
          enum: [scheduled, checked_in, in_room, visit_started, visit_paused, visit_completed, finalization_in_progress, finalized, cancelled, no_show]
        startsAt: { type: string, format: date-time }
        durationMinutes: { type: integer, minimum: 5, maximum: 480 }
        visitType: { type: string }
        modality: { type: string, enum: [in_person, telehealth, phone] }
        source: { type: string, enum: [standalone, ehr_import, clinicos] }
        reasonForVisit: { type: string }
        clinicLocationId: { type: string }
        clinicLocationLabel: { type: string }
        roomId: { type: string }
        roomLabel: { type: string }
        virtualVisitStatus: { type: string, enum: [not_applicable, metadata_ready, link_not_configured, disabled_no_phi_portal] }
        mode: { type: string, enum: [standalone, clinicos_integrated] }
    CreateAppointmentRequest:
      type: object
      required: [safePatientId, clinicianId, visitType, startsAt, durationMinutes, modality]
      properties:
        safePatientId: { type: string }
        clinicianId: { type: string }
        visitType: { type: string }
        startsAt: { type: string, format: date-time }
        durationMinutes: { type: integer, minimum: 5, maximum: 480 }
        modality: { type: string, enum: [in_person, telehealth, phone] }
        reasonForVisit: { type: string }
        clinicLocationId: { type: string }
        clinicLocationLabel: { type: string }
        roomId: { type: string }
        roomLabel: { type: string }
    UpdateAppointmentRequest:
      type: object
      properties:
        clinicianId: { type: string }
        visitType: { type: string }
        startsAt: { type: string, format: date-time }
        durationMinutes: { type: integer, minimum: 5, maximum: 480 }
        modality: { type: string, enum: [in_person, telehealth, phone] }
        reasonForVisit: { type: string }
        clinicLocationId: { type: string }
        clinicLocationLabel: { type: string }
        roomId: { type: string }
        roomLabel: { type: string }
    AppointmentStatusActionRequest:
      type: object
      required: [action]
      properties:
        action: { type: string, enum: [check_in, cancel, mark_no_show] }
        reason: { type: string }
...
    CreateStandalonePatientRequest:
      type: object
      required: [safePatientId]
      properties:
        safePatientId: { type: string }
        displayLabel: { type: string }
        preferredModality: { type: string, enum: [in_person, telehealth, phone] }
    UpdateStandalonePatientRequest:
      type: object
      properties:
        displayLabel: { type: string }
        status: { type: string, enum: [active, inactive] }
        preferredModality: { type: string, enum: [in_person, telehealth, phone] }
...
    WorkspaceValidationResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [workspaceValidation, auditEvent, domainEvents]
          properties:
            workspaceValidation:
              $ref: '#/components/schemas/WorkspaceValidation'
            auditEvent:
              $ref: '#/components/schemas/AuditEvent'
            domainEvents:
              type: array
              items:
                $ref: '#/components/schemas/AuraNoteEvent'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ChartIntakeStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [not_started, metadata_pending, metadata_ready, stale_warning, failed, disabled_phi_upload] }
        sourceFreshness: { type: string, enum: [current_visit, recent, historical, unknown] }
        warning: { type: string }
    ChartIntakeStatusResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [chartContextSnapshot, scheduleMetadata, auditEvent, domainEvents]
          properties:
            chartContextSnapshot:
              $ref: '#/components/schemas/StandaloneChartContextSnapshot'
            scheduleMetadata:
              $ref: '#/components/schemas/ScheduleAppointmentMetadata'
            auditEvent:
              $ref: '#/components/schemas/AuditEvent'
            domainEvents:
              type: array
              items:
                $ref: '#/components/schemas/AuraNoteEvent'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...
    ScheduleView:
      type: object
      required: [appointments, ehrSchedulingEnabled, clinicOsSchedulingEnabled, query, filters, disabledLiveSchedulingSources, metadataOnlyChartIntake]
      properties:
        appointments:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleAppointment'
        ehrSchedulingEnabled: { type: boolean }
        clinicOsSchedulingEnabled: { type: boolean }
        viewMode: { type: string, enum: [day, week] }
        activeDate: { type: string, format: date }
        query:
          $ref: '#/components/schemas/ScheduleQuery'
        filters:
          $ref: '#/components/schemas/ScheduleFilterSet'
        disabledLiveSchedulingSources:
          type: array
          items: { type: string }
        metadataOnlyChartIntake: { type: boolean, enum: [true] }
    ScheduleViewResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/ScheduleView'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    CreateAppointmentResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [appointment, note, patient, linkages, chartContextSnapshot, auditEvent, domainEvents]
          properties:
            appointment:
              $ref: '#/components/schemas/Appointment'
            note:
              $ref: '#/components/schemas/Note'
            patient:
              $ref: '#/components/schemas/StandalonePatient'
            linkages:
              type: array
              items:
                $ref: '#/components/schemas/StandalonePatientLinkage'
            chartContextSnapshot:
              $ref: '#/components/schemas/StandaloneChartContextSnapshot'
            auditEvent:
              $ref: '#/components/schemas/AuditEvent'
            domainEvents:
              type: array
              items:
                $ref: '#/components/schemas/AuraNoteEvent'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    AppointmentActionResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [appointment, note, patient, linkages, chartContextSnapshot, auditEvent, domainEvents]
          properties:
            appointment: { $ref: '#/components/schemas/Appointment' }
            note: { $ref: '#/components/schemas/Note' }
            patient: { $ref: '#/components/schemas/StandalonePatient' }
            linkages:
              type: array
              items: { $ref: '#/components/schemas/StandalonePatientLinkage' }
            chartContextSnapshot: { $ref: '#/components/schemas/StandaloneChartContextSnapshot' }
            auditEvent: { $ref: '#/components/schemas/AuditEvent' }
            domainEvents:
              type: array
              items: { $ref: '#/components/schemas/AuraNoteEvent' }
        meta: { $ref: '#/components/schemas/ApiMeta' }
    StandalonePatientSearchResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [patients, demoFixtureState, realPhiExcluded]
          properties:
            patients:
              type: array
              items: { $ref: '#/components/schemas/StandalonePatient' }
            demoFixtureState: { type: boolean }
            realPhiExcluded: { type: boolean }
        meta: { $ref: '#/components/schemas/ApiMeta' }
    StandalonePatientResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [patient, linkages, auditEvent, domainEvents]
          properties:
            patient: { $ref: '#/components/schemas/StandalonePatient' }
            linkages:
              type: array
              items: { $ref: '#/components/schemas/StandalonePatientLinkage' }
            auditEvent: { $ref: '#/components/schemas/AuditEvent' }
            domainEvents:
              type: array
              items: { $ref: '#/components/schemas/AuraNoteEvent' }
        meta: { $ref: '#/components/schemas/ApiMeta' }
    StandaloneChartContextResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [chartContextSnapshot, auditEvent, domainEvents]
          properties:
            chartContextSnapshot: { $ref: '#/components/schemas/StandaloneChartContextSnapshot' }
            auditEvent: { $ref: '#/components/schemas/AuditEvent' }
            domainEvents:
              type: array
              items: { $ref: '#/components/schemas/AuraNoteEvent' }
        meta: { $ref: '#/components/schemas/ApiMeta' }
    StartVisitResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [appointment, note, visitSession, auditEvent, domainEvents]
          properties:
            appointment:
              $ref: '#/components/schemas/Appointment'
            note:
              $ref: '#/components/schemas/Note'
            visitSession:
              $ref: '#/components/schemas/VisitSession'
            rawAudioRetention:
              $ref: '#/components/schemas/RawAudioRetentionMetadata'
            auditEvent:
              $ref: '#/components/schemas/AuditEvent'
            domainEvents:
              type: array
              items:
                $ref: '#/components/schemas/AuraNoteEvent'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...
    UpdateOperationalTaskRequest:
      type: object
      required: [adjudicationStatus]
      properties:
        adjudicationStatus: { type: string, enum: [open, answered, closed, assigned, deferred] }
        blocksSigning: { type: boolean }
        resolutionNote: { type: string }
...
    UpdateBillingReviewRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [triggered, in_review, needs_clinician, cleared] }
        requestTranscriptAccess: { type: boolean }
        noteForClinician: { type: string }
...
    UpdateIntegrationRequest:
      type: object
      required: [status, reason]
      properties:
        status: { type: string, enum: [disabled, mock_ready] }
        reason: { type: string }
...
    CreateTemplateRequest:
      type: object
      required: [name, visitType, sections, variables]
      properties:
        name: { type: string }
        visitType: { type: string }
        sections: { type: array, items: { type: string } }
        variables: { type: array, items: { type: string } }
    UpdateDotPhraseRequest:
      type: object
      required: [expansion]
      properties:
        expansion: { type: string }
        variables: { type: array, items: { type: string } }
        status: { type: string, enum: [draft, active, retired] }
...
    UpdateEstimateConfigurationRequest:
      type: object
      required: [internalEstimatesEnabled, caveatText]
      properties:
        internalEstimatesEnabled: { type: boolean }
        patientFacingEstimatesEnabled: { type: boolean }
        caveatText: { type: string }
...
    PublishRulesCatalogRequest:
      type: object
      required: [ruleIds, attestation]
      properties:
        ruleIds: { type: array, items: { type: string } }
        attestation: { type: string }
    OperationalActionResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            task:
              $ref: '#/components/schemas/OperationalTask'
            billingReview:
              $ref: '#/components/schemas/BillingReviewQueueItem'
            template:
              $ref: '#/components/schemas/TemplateManagement'
            dotPhrase:
              $ref: '#/components/schemas/DotPhrase'
            estimateConfig:
              $ref: '#/components/schemas/EstimateConfiguration'
            rulesCatalog:
              $ref: '#/components/schemas/RulesCatalogView'
            settings:
              $ref: '#/components/schemas/SettingsAdminView'
...
    TemplatesResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TemplatesView'
    EstimateConfigurationResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/EstimateConfiguration'
...